requirements.txt
python-telegram-bot==13.15
python-dotenv==1.1.0
urllib3==1.26.8
vk_api==11.9.9
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-api-keys==0.5.16
google-cloud-dialogflow==2.41.1